requirements.txt
pandas==2.2.2
numpy==1.26.4
gspread==6.1.2
gspread-dataframe==4.0.0
google-auth==2.29.0
